requirements.txt
Pillow==4.2.1
Jinja2==2.9.6
CairoSVG==2.0.3
docopt==0.6.2
fabulous==0.3.0
pytest==3.2.1
